// packages/model-registry/src/concepts/modelRegistry/types.ts
export enum ModelState {
  LIVE = 'LIVE',
  ARCHIVED = 'ARCHIVED',
}

export enum InferenceServiceState {
  DEPLOYED = 'DEPLOYED',
  UNDEPLOYED = 'UNDEPLOYED',
}

export enum ModelSourceKind {
  CATALOG = 'catalog',
  KFP = 'kfp',
}
// ...
export type ModelRegistryCustomProperties = Record<string, ModelRegistryCustomProperty>;
export type ModelRegistryStringCustomProperties = Record<string, ModelRegistryCustomPropertyString>;

export type ModelRegistryBase = {
  id: string;
  name: string;
  externalID?: string;
  description?: string;
  createTimeSinceEpoch: string;
  lastUpdateTimeSinceEpoch: string;
  customProperties: ModelRegistryCustomProperties;
};

export type PipelineRunReference = {
  project: string;
  runId: string;
  runName: string;
};

export type ModelSourceProperties = {
  modelSourceKind?: ModelSourceKind;
  modelSourceClass?: string;
  modelSourceGroup?: string;
  modelSourceName?: string;
  modelSourceId?: string;
};

export type ModelArtifact = ModelRegistryBase &
  ModelSourceProperties & {
    uri?: string;
    state?: ModelArtifactState;
    author?: string;
    modelFormatName?: string;
    storageKey?: string;
    storagePath?: string;
    modelFormatVersion?: string;
    serviceAccountName?: string;
    modelLocationType?: ModelLocationType;
    artifactType: string;
  };

export type ModelVersion = ModelRegistryBase & {
  state?: ModelState;
  author?: string;
  registeredModelId: string;
  labels?: string[];
};

export type RegisteredModel = ModelRegistryBase & {
  state?: ModelState;
  owner?: string;
};